requirement.txt
fastapi
uvicorn
requests
python-dotenv
gTTS
pydantic